Initial Commit
This is the day I rewrote half the program from scratch to work like a
real websockets app. With this new structure it can process input no
matter when it comes or what kind of data it is.

Also continued to work on notching system, and got low level notching
functions working well (well enough) on the ESU EMD 567 decoder.