Additions, mainly setReverser()
Added some functions, cleaned up some code, the big addition being
setReverser(). In it's current state, setReverser() can be called with 2
parameters, direction (either "forward", "reverse", or "neutral") and
ignoreNotch, which is used for when JMRI alerts LocoThrottleJS that the
direction of the locomotive has changed (for example if someone with a
normal throttle changes it). Without ignoreNotch set, it defaults to
false, and the reverser will not set it's new value unless the
locomotive is idling (notch == 0). The setReverser() function returns
true if setting reverser value was successful, and false if it was
unsuccesful. It also enters a small entry to console.log()

I also added global-vars.txt for my own reference, mostly as a place to
list variables and jot down notes as to what they are.