Fixed issue with reverser
In the last commit, I didn't notice that setting the reverser (even if
it was the same) sent a new command each time, so I added a "If its
already set to this, just reset the variables and client side stuff,
don't bother sending another command" part. I also changed decoder.js's
JSONhandleType_throttle_functions. I moved the part that handles
incoming reverser info into infohandler.js, so that it's harder for
people to mess that up. Since decoder.js is decoder specific, they'd
have to copy and paste my code for it to work reliably, so might as well
make their lives easier.