Turned locoAddress failsafe system into a function
I decided it'd be far easier if future decoder.js developers had access
to an easy way to reproduce that alert system. So I turned the functions
of the weird code I inserted into things, into a function: sendcmdLoco()
It works JUST LIKE sendcmd(). In fact it actually encapsulates
sendcmd(). It just ALSO has that fancy alert system about not having a
throttle requested, thus the name sendcmdLoco() because it's only
relevant to loco commands.