Significant addition to setNotch() and notching
This is the biggest update to the notching system yet. This adds a
functional setNotch(up or down), with a setNotch(number) in the works.
It also adds the updateNotchHTML() function, which is called every time
the notch is changed. It's in ui.js, so UI developers can write their
own scripts to update the notch indicator. It's pretty much open for
them to do as they please.

ProtoEngine is coming along nicely. If I can just get the notching
system to where I can do setNotch(number), then I can move on to braking
and the speed algorithm.