FIRST DEV RELEASE! Notching system is functional!
The notching system is fully functional! We basically have a
prototypical prime mover notch system, and a prototypical (I think?)
reverser. There "are no brakes on this locomotive". Yet.

Also added a reset prime mover option, which (in my ui.js, at least)
runs when you get the throttle. It makes sure the locomotive is stopped,
the notch is set to 0, and the prime mover is cut off. That way there's
no way to confuse it.

And I also added a few more global variables, including one which gives
the option to change how many notches your loco has. I figure maybe
someone wants to use this with a weird GE 16 notch engine.