Added locoMaxSpeed cfg parameter
I added a parameter (set in the locoMaxSpeed global variable) in case
you have a locomotive with a top speed that's too fast to be
prototypical. In setting this variable, you set (out of 100 speed steps)
what the highest speed you want this loco to go is. The notching and
everything will still behave just as it always has, just the speed
algorithm treats whatever you set here as the top speed instead of 100.
So if I set 80, the highest speed LocoThrottleJS is ever going to make
my train go is 80% throttle, and that's on notch 8 with no brakes
applied.

This feature has been tested to work. It is also decoder agnostic, so no
further testing should be needed.