Locomotive Air Brake functional!
This is the first commit where we've had any kind of braking system. The
locomotive air brake is now functional, and able to be set from HTML.
Its physics are not perfect, but it does work, and the physics can be
adjusted later on. Right now the program doesn't perform any kind of
delayed reaction, lurching, coupler slack reaction, or extra momentum.
It simply takes in parameters and spits out speed.