accel() function added
I added an accel() function that came to me suddenly. There's a comment
line in ProtoEngine.js about how to call it, but basically all I'll have
to do is figure out a way to let the rest of the program know that the
variable has been updated, and that won't be hard.