Added functional reverser for physics engine...wind still broke
The reverser option now functions! When you set the reverser to neutral, the engine outputs 0hp regardless of notch, but when you set it to forward it works just as before. If you set it to backwards, it outputs the same amount of horsepower, but negative instead of positive, thus moving the train in reverse. The wind resistance is still horribly broken for reverse, but forward works great.