Fixed forward/reverse movement physics, wind resistance functional
The forward and reverse motion are now completely functional, or so I think. If anyone else on the interwebz knows physics more than I do, feel free to chime in and tell me what dumb mistake I made :P
I removed the option for natural wind, simply because it's too much to handle right now. If somebody wants to help add an option for wind based on which way it's blowing in relation to the loco, that's fine, but I don't think it will be very realistic since the loco would turn and the wind would change direction.

Wind resistance in a calm environment should be accurately simulated though, the wind the loco feels from it's own motion is all taken care of.

I also tested, just because, what would happen if you get going forward, and then put the loco in reverse. Sure enough, it's like using your brakes. Vice versa and the same is true. So I think my forward/reverse motion is finished, braking is next.