Mostly working physics
The physics engine is mostly working as far as I can tell. I fixed the wind resistance issue, it works with both fwd and rev. AFAIK from my tests. The issue now is that during deceleration, we never reach exactly 0, we always just dance around 0, which throws the reverser every 100ms. I need to figure out an anti-infinite-loop thing so that the reverser detects when it's going to do that, and ignores it and sets the speed to 0...dunno how but I need to figure it out