Major Upgrade: Rewrote physics engine to minimize if/else usage
I changed the reverser system so instead of being a string ("forward", "neutral", or "reverse") it is simply a 0, 1, or -1. This lets me get rid of the clunky if statements to deal with directional nonsense; it's all math now.

I also fixed some boneheaded math errors, for example my rolling resistance was calculated in a way that would only work if Earth's gravity was doubled...not good...so I fixed that and a few other odds and ends.