Began integration of notch slider with setNotch()
I began setting up the HTML <range> slider to play nice with the setNotch() function. I got it where every time a user sets it, it runs setNotch(thevalue), but setNotch must be upgraded to support setting a number notch before I can use that. The slider DOES, however, only allow a value change if setNotch() says it can, so right now it is permanently stuck on 0. This is cool because it means the notch timing limit will actually work without a hitch.