MAJOR FUNCTION NAME CHANGE
I changed the name of the setNotch() function to setNotchCrude(). It still works the same, but will be added to with a new setNotch() function which accepts numbers, and I want the simple name for the one everything needs to use. setNotchCrude() is just for setNotch() to use when it sets the notch.