Fixed bell HTML system, all future functions should be modeled after the bell function in this commit
I finally fixed the infinite loop system that the bell would get into. I
had to do some if/else stuff to make sure it didn't run any of the
update functions if the value wasn't actually any different, and that
kept it from just running them over and over in a loop. Weird, but it
works.