Added debugToast(), finished air horn button
There is a new function now, which can be called with
debugToast(stringfortoast, timeinms)

Toasts called using this function will only display if debugToastMode =
true, so use this for any kind of development toasts or things that
normal users should not see.

Also finished the air horn button, and tested it. It works!