Functional Script Selection System
The script selector is functional. It is very un-user-friendly, and a
little clunky, but it is functional and will be easy enough to expand on
later for better UX. Also, I fixed the layout track power switch so that
it automatically updates based on WebSockets data, much like I fixed the
function switches the other night.