First function train builder, Materialize mods
The train builder GUI is now partially functional, with only the option to add a locomotive.

The changes to MaterializeCSS are because I compiled it from SASS instead of using their provided file. It seemed to work better.

I also modified the JavaScript for Materialize. Now you can make any element inside a chip close the chip by simply giving it the class "close-chip-btn". Icons are not automatically made into close buttons.