Tinkering, committing this before do-over
This is a commit that I may need to revert to later. I tinkered with the
physics engine a lot but realized I may have made a colossal blunder in
calculating motive force. I plan on rewriting this, but I wanted to
commit the changes here so I have something to revert back to.