First support for tractive effort curve
This is the first commit where we've supported the equation for building
a tractive effort curve from Virginia Tech. The maximum speed parameter
is not set up yet, but best I can tell the motive force calculation
system is fixed. Not sure about wind resistance yet, still looking into
that.