Major UI updates for gauges
Using Google's gauges library (from their charts system stuff) I added most of the beginnings of true dashboard gauges. Much of the fancy details and whatnot is still in the works but the basic concept is down. In this commit I outlined the syntax of the gauges in prototypes.json. In every locomotive object, there is a sub-object "gauges" with sub objects for each gauge. So for example, train[0].gauges.speedometer would give properties of the speedometer. This may be subject to change (specifically what options are present for each gauge) but the basic structure should remain the same.