THIS IS A VERY UNSTABLE COMMIT.
I worked on this in LIS class. I am trying to redo most of the lower level
core JMRI/DCC stuff to better support MU and other things. I want to have
decoder functions integrated into the appropriate train variable.

The goal of this commit is to be able to call:
train[0].decoder.f.bell(true);
and have the bell come on on your lead locomotive.

That was FAR from achieved, but I am committing the changes just the same.
Perhaps I can reuse this code/fiddling later.