LANDMARK COMMIT: object-based throttle system
This marks the completion (of the first iteration) of the new object-based throttle system.

To create a throttle, you can call someObject = new jmri.throttle(1379, "myThrottle"). That will create a throttle called "myThrottle" for address #1379.

The plan is to use this with the train object, with the train position number as the name. So train[0].throttle.f.set(0,true) would set F0 to ON on whatever locomotive train[0] is. This is not implemented yet, but the underlying throttle object system based around the "new" keyword is totally done (best I can tell).