Backstage decoder object done
The framework for the new decoders object is finished (or mostly so).

The object is based around a sorting system that sorts like so:

manufacturer>product>sound project

This means that, for example, a LokSound Select decoder with an EMD 567 sound file on it would have the following:

decoders.esu.loksoundSelect.emd567

There will need to be an approval process for integrating a developer's decoder file into the "central" object (stored in the GitHub for everyone to download), but it should ease things a bit.

The way that this is used is simple. This new system integrates with the new JMRI/WebSockets system, explained in the previous commit. It acquires a throttle when you create the dcc object. To set up the first element in the train object as a LokSound Select decoder with an EMD 567 sound file, on DCC address #2, you would do the following:

train[0].dcc = new decoders.esu.loksoundSelect.emd567(1379, 0)

The 1379 is obviously the locomotive address, but the 0 is important. That must always be the same number as the train position. So train[1] would have 1 in place of the 0. This is the system I have developed for essentially retro-fitting JavaScript with the ability to see parent objects. If the dcc stuff knows it's on train[0], it can use that information to get access to everything else it needs to know. It is quite important to note that the throttle object is integrated with the dcc decoder adding function. When you add the decoder information with the new keyword, you also automatically add the necessary throttle information, contact the JMRI PC, and request the throttle for the locomotive. Simple!

This system is obviously still WIP status, but the underlying structure (at its most basic level) should be pretty much unchanging from here on out. As of this commit, none of the userspace/UI stuff to make this work is done yet. This new system can ONLY be used from the console, however upon testing it does work. I can confirm there are no variable conflicts or issues with multiple locomotives, as I tested it with 2 locomotives simultaneously and I was able to independently control each one.