First steps on mockup of improved train builder
I have begin working on the mockup of the new train builder UI. I say
mockup because it is not a functional UI, it is only the HTML/CSS to
make it work. I have not written any JS to make it automatic yet, I am
only laying out how I want it to look.

As of this commit, I have determined that I will keep the existing
"chip" adding system, but I am going to add a link to the text in the
chips. This link will open a modal (with a fixed footer) where users
will be able to see information about the locomotive (or rolling stock
element) and configure things like DCC decoder object, DCC address, and
the model object. They may also eventually be able to create a package
file that bundles all this info together and is stored in cookies, but I
don't know. I'll have to look into it more.