Mostly finished mockup for train builder
I have almost completed the mockup of the train builder UI. I used Materialize's "modal" option to make the chips clickable. I am planning to begin working on the JavaScript to make it all functional tomorrow, but this commit contains a mockup of the UI that I will be using as a reference.