Finished more core roster functions
This commit is pretty major. I added the following variables:

jmri.roster.raw - This is the raw version of the JMRI roster. This is stored EXACTLY as it is returned via WebSockets from JMRI.

jmri.roster.entries - This is a much more dev-friendly version of the roster, in my humble opinion. This has the roster organized by entry names, with the base object having the names of the entries, each with their data object as the value. So you could call jmri.roster.entries["GP15-1 #1379"].decoderFamily and get the decoder family for your geep. It's simpler and easier in my opinion. It also stays automatically up to date with the raw roster, thanks to some changes I made in websockets.js

I also added the following functions:

jmri.roster.reformat - this function is only used by backend scripts, but it takes the raw roster and builds the jmri.roster.entries variable.

jmri.roster.matchProperty - Per the docs in the code: This function is used to find entries in the JMRI roster which match a certain object.

You call it with:
jmri.roster.matchProperty({"property":"value"})

The function returns the name keys of all the entries that have the property with the correct value.

For example, jmri.roster.matchProperty({"decoderFamily":"fakeDecoderFamily"}) would return an array of the names of every locomotive whose decoderFamily attribute equals "fakeDecoderFamily". If nothing fits the query, it will return an empty array, or [].

I worked some on the modal HTML builder, but this is not a commit for that. That code is still highly unstable and has a significant chance of changing to be more roster-based. This commit is PURELY for the roster-related update.