Worked on train builder and train objects
I began work on the new improved train builder UI. I was able to get an add function going that works using a bundle object.

The biggest, most significant change of this commit is that the bundles.json object system is standardized. Everything will be based around that, with a "developer mode" option that opens up a textarea for you to type in your own bundle file if you wanna do crazy stuff. Since I haven't gotten any votes from ModelRailroadForums.com, I went with this because I like this option best myself.