Added slate /docs folder
I added the folder where the Slate-based documentation will go. Since it's all static HTML, I'd rather have the version control tools of GitHub pages than host it on my own site, so I'm putting it here for now. This is an unfinished documentation set obviously, but it's a proof of concept.