Finished first version of new landing page
The first new landing page is done now. I haven't added the final stuff to make it easy for newbies to get started (because we're not to that point yet) but it is a functional system.

It contains some custom JS work I did for transitions, as well as two third-party JavaScript libaries.