Finished notching slider overhaul
The notching slider's scripting work is done. It will always reset itself to the actual notch value, so if the user tries to move it farther than 1, it will just reset to the actual notch value (determined by notch.set()). This is more efficient than the conglomerated approach I was using.