Rewrote engine startup checkbox
The engine startup checkbox uses the new object-based UI system now. It is not actually integrated with the throttle system yet, as I am trying to decide the best way to do that, but I have gotten it ready to integrate with a method of my choice.

ui.cab.engine.start
  - set() : the function, which accepts true or false and automatically updates the checkbox (or forces it back to the previous position)
  - state : a boolean