Rewrote headlight switch to be object-based
Now:

ui.cab.headlight.set(true/false)

Is used to set the headlight. This function is called using onChange of the checkbox. The checkbox ID is "ui.cab.headlight" now. The function, when called, will make sure the checkbox reflects the state correctly. Most of the time this means basically nothing, but in instances where the lead locomotive doesn't exist, it will not allow the user to turn the headlight on.