Fixed TypeError for recieving roster data
Previously, the WebSockets function ``processReply()`` would throw a typeerror about roster stuff because it was trying to read a property of a list element. If the list element wasn't defined (so if it wasn't a roster) then it would just freak out and throw a TypeError. This commit fixes that; I added a thing to check if the data is a list at all. So if it's not a list, we just skip the whole mess, thus avoiding the TypeError.