Finished new gauges system
The new gauges standard is done.

It makes use of this excellent library: https://github.com/Mikhus/canv-gauge/

ALL the options for the gauges are specified in .prototype.gauges.<gauge name> as an object, just like the gauges library needs them. Developers have access to EVERY option except renderTo. This leaves the door for styling/realism open to just about anything.
Gauges are updated with gauge.<gauge name>(value)

So for example, one might call:
```javascript
gauge.rpm(300)
```
which would set the RPM gauge to 300.

I have completed a proof of concept for this new system as of this commit; the speedometer and the RPM gauge are proven to be functional.