Added example sound notch system to decoder example
I reused some of the implementation in the original loksound.js (written in June 2015) to add support for notching.

There are 3 key things inside the .dcc.f.notch object:
- up() {function}
- down() {function}
- state {number}

Up and down should move the notch up or down ONE CLICK. It is important to remember to add protection against notching below 0 or above 8; I did that with a couple IF statements.

State is the current notch of that decoder. 0 is IDLE, 8 is RUN8.

The implementation I put in the example decoder object has been tested to work with an ESU LokSound Select decoder with the EMD 567 12cyl. sound file.