Fixed massive bug in train.build.remove()
As it turns out, train.build.remove() wasn't working nearly how it should've. It was messing things up big time by not removing the correct item from the train, and sometimes not removing things at all, which often duplicated things in other places. It was a mess.

The reason for this bug was actually simple: I was trying to use indexOf() in a place where it won't work.

Since train.all is not an array of strings, but is instead an array of objects, you can't use indexOf to find the array position of a train element with a given name. You have to find it by looking through every element's roster.name property. To do this, I added a simple function:

train.all.find(name)

Arguments:

- name - This is a string to search for. This should be the roster.name attribute that you want to find.

Returns:

position - The function returns a single number, which is the position of the item in the array. If the item name you searched for doesn't exist, it will return undefined. Simple!

This function has been added to train.build.remove() to replace .indexOf(), so as far as I know (I tested it briefly) the train builder actually works correctly now.