Fixed fuel consumption rate with engine off, added switchable cabs
I did a quick bug fix to make the fuel consumption rate automatically be 0 when the engine isn't running. This also means the tractive effort is zero when the engine isn't running.

I also finished the low-level functionality to switch cabs. To switch, you call:

ui.cab.setCurrentLoco(name)

name : this is a the locomotive's roster name.

This function will automatically do everything needed to switch the cab.