Updated docs more, regenerated static HTML
I updated the docs in a number of ways. The most significant updates were to:

Standards>Prototype
Standards>Decoder

I added GitHub Gists in place of the previous examples in both of those. This should allow easier maintenance of these standards, and makes it easy for developers to participate in maintaining them.