Worked on compressor logic more, adjusted prototype standard
I have worked more on the compressor's logic. I have not finished the pressure calculation system yet, but I have finished a system for calculating flow rate, as well as logic for turning the compressor on and off based on pressure of main reservoir.

The improvements to the standard should be documented shortly, but involve exclusively air-related objects.