More work on air compressor stuff
I have decided (at least for the time being) to use a constant consumption rate (like in cfm) to remove air from the reservoir(s). This will be set in .prototype, and will probably be in cubic-feet/millisecond so I don't have to do any conversion math in the software.

In this commit I worked on the beginning of the horn air consumption. it is not stable yet so don't try it.

I also fixed some subtle little compressor bugs, such as the sound not turning off when you kill the engine. The logic for filling/maintaining main reservoir pressure is, best I can tell, totally done now.

Merry Christmas!