First iteration of horn air usage system
I have completed what appears to be a mathematically sound horn air consumption system. It uses an interval at every 1ms to take away a very tiny amount of air.

I also fixed a weird issue with the main reservoir pressure gauge where it would have a spasm seizure whenever you updated the value. It was checking to see if the value had changed before it actually updated it, but the problem was it was checking it against another gauge's value. So yeah, that was fun, because it took 30 minutes to figure it out and all it ended up being was that I couldn't read.