Worked more on air logic, half-finished air dump
I worked a lot more on the air systems logic. I began work on the air dumping system, and I got a lot of it done. Working features include:
compressor won't start when air dump valve is open
sound works on CBQ2 LokSound decoder profile
switch in cab partially works

Still-flaky:
auto-shutoff of functions that use air (namely the bell, and in the future, the brakes)