Fixed excessive redundant setting of compressor
Until now, the compressor state would be set on every physics sim cycle, regardless of whether the state actually changed. This resulted in a ridiculous amount of redundant WebSockets (and consequently, DCC) packets. Some command stations may have problems with this (my Digitax Zephyr did on occasion) so I deemed it necessary to fix. Now ZephyrCab only sends a command when the state is actually changing, as opposed to every cycle.