Fixed unhandled exceptions due to unreleased throttle
Previously, due to a combination of my poor research and what was a poorly documented feature of JMRI, ZephyrCab would never actually release the throttles. Even when the connection was closed, JMRI held on to the throttle info. So whenever ZephyrCab would relaunch, it would start at "1" for the throttle name, and there would be a big naming conflict. JMRI, for whatever reason, really hates this, and it throws a lovely unhandled exception to the console and promptly kills the WebSockets connection.

This is no longer an issue, as I have implemented a simple function built into .throttle that releases the throttle to bring JMRI up to speed on what we're doing. As of this commit, I can no longer reproduce the aforementioned bug, so I am assuming this commit fixes it.

Function: train.all[x].throttle.release() - This will release the throttle. It doesn't take any arguments, and it is automatically called by train.build.remove upon taking an item off the train.