Implementing new Net Force code
There is new code towards the bottom of sim.js's main sim.accel loop. For simplicity's sake, I decided to calculate the acceleration of the train as one full unit by treating it as one big mass and finding one big net force. This method may prove too rudimentary later on, but for now I see no issues with it, especially since we're still left with the ability to pull more localized numbers from individual train objects (for example, train.all[0].prototype.realtime.te is still going to exist. If we need information on some specific element in the train, we're not totally screwed with this new method,  it just makes the main function of the sim.accel function much easier.