Finished computation of total train weight and total train net force
I added some one-liners to add train.all[i].prototype.whatever for each element together to make train.total.whatever. These whatever's, as of this commit, include only weight and net force.

One perk to this approach is that grade force will be very easy, since it's calculated for each element based on it's weight and grade. This means we don't have to fight with that on a large scale, just toss every value into the train.total.netForce value and it'll all work out. Like magic!