Finished first acceleration system
After fixing a variety of variable issues, including the syntax of the zero-out lines at the beginning of sim.js, I have a functional acceleration system that accounts for TE, rolling resistance, and grade resistance.

The speedometer on the dash does not work in this commit. There is also a lot of debug information spit out on the console, so it is unintuitive to debug manually due to the lag of the console.