Fixing acceleration units, enabling speedometer, adding forced zero crossing
In the last commit, the logic was correct for acceleration, but we were supposed to use kilograms (not grams). This meant everything accelerated 1000x too slow. That issue has been fixed in this commit.

I also implemented the speedometer in sim.js, which led me to notice that the old forced-zero-crossing logic was no longer there. I re-implemented that and confirmed that the software is functional.

There is no actual model reaction to the acceleration (yet), that's coming later. I have to first create a function mapping mph to DCC speed steps, which will be a challenge without a speedometer.