Improving comments, adding maxSpeed calc func
The main focus with this commit was the creation of a
.prototype.calc.maxSpeed function. The idea of this function (also
documented in the comments) is to take the trainPosition argument for an
element and then return the maximum speed it can reach. This is the
beginning of my addressing #2 , and leaves the theoretical stuff
relatively open since there will probably be people with several ways
they'd like to handle this.

Other than that, I just went through bundles.json and improved some of
the commenting.