Adding basic implementation of element-specific max speed
In this commit, I added the first implementation of the maximum-speed
tractive effort limit system as described in #2 . The implementation
seems sound in principle, but it has not been tested and therefore it's
likely that there are programming errors. I hope to test later tonight
and follow with another commit and closing #2 , tying up loose ends, and
adding a little better comments to the associated functions.