Completed new .throttle.f.set()
The new function should address the issue of multiple functions not being triggerable at the same time. Instead of accepting only 2 arguments (a state and a function number) like the old function, this new one accepts a JSON array as its only argument.

For example, .throttle.f.set({"F2":true, "F6":false}) sets F2 to true and F6 to false, leaving all the other functions untouched. The function will accept anywhere from one argument to all 28 functions, and any functions that are not defined in the input arguments are not included in the outgoing WebSockets packet.

I tested this function on a low level with JMRI, and verified it works. I have not updated the DCC decoder control objects to use this new architecture. That will come in a later commit.