Introducing "force" option for .dcc.f air devices, fixes #5
I introduced an optional argument for the air devices in .dcc.f. for my template LokSound decoder object. When not specified, the function works the same as it did before, but when the "force" argument is set to true, the function will disregard any restrictions on action based on air pressure. Consequently, it's no longer necessary to perform the post-air-dump shutoff in a certain order, since we can force the functions to shut off regardless of the fact that there is zero air pressure.

In brief testing, this resolved #5, but it may need to be reopened again at a later date.