Commenting out temporary console.error in jmri-core.js
There is a console.error line in jmri-core.js that, when uncommented, will show every outgoing DCC function command as a console error. This is useful for debugging since it makes it very easy to filter, but not necessary in the long run, so I commented it out until it's needed again.