Resolved #11, worked some on beginnings of wheel slip math
The main focus of this commit is the former: issue #11 has been resolved. This issue was a TypeError that occurred when the startup switch was flipped (in either direction). Upon further testing, somehow the variable i (for the FOR loop in sim.accel) was getting to be 29, regardless of train.all.length. I'm not sure where the 29 was coming from, but by prefixing the initial i=0 with the var keyword to define it only in that scope, I solved the problem. It's a peculiar solution (at least I think so) but I tested it thoroughly and it does solve the problem. See the issue page for more information.