Writing beginnings of rolling stock sim.accel routine, fixing obscure sim.accel bugs
I started writing the calculations in sim.accel for rolling stock. This is pretty straightforward since there is no motive force, only resistance, and in this commit I have not even  started on brakes yet, so it's all just constant. See the comments in sim.accel on this commit for more details on that; nothing spectacular though.

On the other hand, some pretty major moving-of-stuff was done in sim.accel's existing code. While writing the rolling stock math, I discovered that all the f=ma calculations re-ran at the end of EVERY cycle of the main FOR loop in sim.accel(). This meant that false speed values were calculated because it had not had a chance to add up all the values for train.total yet. I moved the appropriate code outside the FOR loop, which revealed another bug.

The instant sim.js is loaded, there is a setInterval() that's run to start the sim.accel() cycle. This happens LONG before the user has a chance to build their train. This wasn't noticed (I'm stupid) previously because if the train.all array was empty, the main FOR loop didn't run even once, which meant no math was done. However, upon fixing the bug above, since I moved code OUTSIDE that FOR loop for the first time, things went downhill. Since that code runs regardless of train.all being empty or not, it was trying to do math with variables that didn't exist, which meant it set the speed (and lots of other things) to NaN. The solution to this problem sounds simple: make the whole function exit before doing anything if train.all is empty. That's exactly what I did, it's just a weird, wonky solution. See below:

if (train.all.length == 0) {
        return null;
}

Nothing checks the return value of this function, but the return statement ends the function before it has a chance to do any math. This  code can be found on line 58 in this commit.

I have tested everything and it seems to work fine, however it's not improbable there are a few more problems. This commit should be stable, but stable is a relative term.