Working on brake logic, equalizing pressure math
I worked on the structural component of the brake system, and made a few variables and some empty functions in scripts/brakes.js . The big step forward for this commit is the addition of the brake.findEQpressure() function. Called with one argument, the feed valve setting, it calculates the maximum possible full service reduction as well as the equalizing pressure. It returns them as an object like the one below:

For a 130psi brake pipe:
{
  EQpressure: 92,
  fullServiceReduction: 37,
}

The algebra/logic behind this function is explained in a VERY long comment inside the function.