Resolving #13 (object cloning bug)
I solved the problem with a small modification to train.build.add, which means no significant rewriting is required. I can go on about my business now, with objects that don't refer to each other, and without having to rewrite huge chunks of established code.

See Issue 13 on GitHub for more details on the problem.