Work on train brake pipe and timing system
The primary purpose of this commit is the timing system through which pressure changes are propagated. Each car can take a given number of milliseconds (set in .prototype.brake.latency) to propagate the signal, prototypically accurate and realistic behavior. This was somewhat difficult to keep responsive since most cars take 100ms to propagate, but I settled on the following:

- ``brake.cycle(car)`` takes a single argument and sets up the necessary PSI changes (using setTimeout) for the train.all element specified in the argument
- ``sim.js``'s rollingstock section has a ``for`` loop that runs ``brake.cycle()`` across the whole train for every one rolling stock cycle. Since the function doesn't take much time to actually execute, it doesn't have a big impact on ``sim.accel``'s performance, but it also makes sure every car gets re-"parsed" at least every 100ms (or however often sim.accel runs).

**Variables Added:**
- boolean .prototype.brake.waitingOnChange - This is ``true`` whenever a ``setTimeout`` is already configured for a car. This keeps ``brake.cycle()`` from setting up multiple ``setTimeout``s on a single car for the same pressure change.
- number .prototype.brake.latency - This is a time, in milliseconds, that it takes for a pressure change to propagate through a car.

I have tested the timing system, which alerts to the PSI changes using ``Materialize.toast()``, and it seems to work reliably. This is not a finished braking system, and particularly it lacks a lot of the special features needed in locomotives (such as bail-off for the independent brake handle), but it is a start.