Quick improvements to sim.start() and sim.stop()
I added some finishing touches on these  two functions. sim.start() stores the time you feed to it in sim.time.interval, and if you don't specify a time it uses the last time stored there (100 by default). sim.stop() hasn't been modified, but the changes to sim.start() make it much easier to pause/play the simulation. The purpose of this was to implement a brief pause for a brake system reset, code for which has not been committed yet.