Adding automatic brake charging/setup for new elements
With this commit, when a new element is added, ``brake.fixNewElement()`` is called with that element's position in ``train.all``. This function automatically configures the brakes correctly as though the reservoir/pipe are charged. Technically, the pipe would not be charged when coupling a lone car in real life, but ain't nobody got time to wait on that. The reservoir _would_ be charged, so that's accurate.

I tested this new system and confirmed it works. It is worth noting it makes use of  ``sim.stop()`` and ``sim.start()``, but I double checked to make sure and it does start the sim up properly when it's done.