Adding "average" train line pressure calculator for gauge
I added a simple function, ``avgLinePSI``, which will calculate the average of all the ``.prototype.brake.linePSI`` values on the train. This is the value I plan on showing on the train brake pipe gauge in the cab, not the localized locomotive value. I have not tested this on a fully functional system yet, but it works fine with a makeshift test system and JavaScript doesn't find any syntax errors.