Fixing .reduction() to work with multiple consecutive reductions (resolving #15)
Previously .reduction(), unbeknownst to me, would work fine for single reductions. But if you made another reduction, it would start at the feed valve pressure instead of at the actual current pressure. See #15 for more info on this. This commit fixes that with a simple change to a few lines, and it resolves #15.